Fix quaternion left-multiplication as global rotation
